- Added Jolokia support to the monitor - Disabled the network controller out-of-the-box as it was continually updating a bridge after adding a consumer to   one of two brokers networked together. - More work on the service manager.   - Renamed artifacts and directories from server-xxx to service-xxx. - Registry utility class with registry listener and concurrent map handling.

+++ amq-bridge/src/main/java/com/amlinv/activemq/monitor/model/ActiveMQBrokerStats.java
@@ -16,7 +16,7 @@ public class ActiveMQBrokerStats implements MBeanLocationParameterSource {
 
     private long memoryLimit;
 
-    private int memoryPercentUsage;
+    private long memoryPercentUsage;
 
     public ActiveMQBrokerStats(String brokerName) {
         this.brokerName = brokerName;
@@ -53,12 +53,12 @@ public class ActiveMQBrokerStats implements MBeanLocationParameterSource {
         this.memoryLimit = memoryLimit;
     }
 
-    public int getMemoryPercentUsage() {
+    public long getMemoryPercentUsage() {
         return memoryPercentUsage;
     }
 
     @MBeanAttribute(name = "MemoryPercentUsage", type = int.class)
-    public void setMemoryPercentUsage(int memoryPercentUsage) {
+    public void setMemoryPercentUsage(long memoryPercentUsage) {
         this.memoryPercentUsage = memoryPercentUsage;
     }
